fixed edit product page
