+++ src/App.jsx
@@ -18,7 +18,8 @@ import ProtectedRoute from "./components/ProtectedRoute"
 import SuccessPage from "./pages/SuccessPage"
 import CancelPage from "./pages/CancelPage"
 import AuthOutlet from '@auth-kit/react-router/AuthOutlet'
-import Products from "./components/Products"
+// import Products from "./components/Products"
+import ProductPage from './pages/ProductPage'
 
 
 function App() {
@@ -48,7 +49,7 @@ function App() {
           </Route>
           <Route path="/api/dashboard" element={<ProtectedRoute><Dashboard /></ProtectedRoute>} >
             <Route index element={<ProtectedRoute><Stats /></ProtectedRoute>} />
-            <Route path="products" element={<ProtectedRoute><Products /></ProtectedRoute>} />
+            <Route path="products" element={<ProtectedRoute><ProductPage /></ProtectedRoute>} />
             <Route path="orders" element={<ProtectedRoute><Orders /></ProtectedRoute>} />
           </Route>
         </Route>
